This is the initial online audio TCP server. It accepts clients on a TCP socket and reads raw audio (16bit 16 kHz signed PCM) in a format where each packet is prepended by a 4-byte size. A packet of size zero (that is 4 zero bytes) denotes the end of the stream and prompts the decoder to output the result.
The result is sent on the same socket as the audio immediately after the audio
is received. The format has a simple human-readable header and one word per
line output. This format can be expanded as pointed by the "format" argument
in the header.

So far, everything is contained in a single file, to simplify deployment, but
this can be changed in the future.