sandbox/sail: Fixes for Mac OS 10.9 specific makefile.
This commit adds support for vanilla gcc (not the gcc binary that comes with
Xcode command line tools which is clang in disguise) on Mac OS 10.9.
Specifically:

1. Adds -stdlib=libstdc++ option only when compiling with clang since vanilla
gcc does not recognize this option.

2. Removes the -rdynamic option from src/makefiles/darwin_10_9.mk since it was
causing an undefined option error with vanilla gcc on Mac OS 10.9. This worked
in the past because Apple's gcc compiler was silently ignoring the -rdynamic
option (it is for ELF binaries only). Apparently the behavior enabled by this
option, adding all symbols to the dynamic symbol table, is the default
behavior on Darwin.

3. Adds -flax-vector-conversions option when compiling with vanilla gcc. This
tells recent versions of gcc to allow vector conversions without an explicit
cast provided the vectors are of the same size.

4. Defines CC to be the same compiler as CXX to avoid different parts of the
codebase being linked with different C++ standard libraries when CXX
definition is overridden by the environment but CC is not. This is again
related to clang linking with libc++ by default.

5. Adds -Wno-unused-local-typedefs option to silence the annoying typedef
warnings originating from openfst.